requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.21.0
Pillow>=9.0.0
google-generativeai>=0.3.0